refactor: drop trackedDialog rename, fold analytics into dialog(item:)
The trackedDialog modifier was over-engineering. DialogItem.tracked
already encodes the analytics decision via the factory choice; repeating
that intent in the modifier name added call-site noise without
conveying new information.

Moves the DialogItem-specific .dialog(item:) overload from FlipcashUI
to the Flipcash app target so the modifier can dispatch
Analytics.errorModalDisplayed directly. FlipcashUI keeps its general
.dialog(isPresented:) and .dialog<T>(item:style:...) overloads. The
DialogItem rendering primitives (PartialSheet, Dialog, Dialog.Style)
were already public; backgroundColor is now public too so the
app-side modifier can resolve it.
