+++ Flipcash/Core/Screens/Onramp/ConfirmPhoneScreen.swift
@@ -95,7 +95,7 @@ struct ConfirmPhoneScreen: View {
             .padding(20)
             .foregroundStyle(.textMain)
         }
-        .trackedDialog(item: $viewModel.dialogItem)
+        .dialog(item: $viewModel.dialogItem)
         .navigationTitle("Verify Phone Number")
         .toolbarTitleDisplayMode(.inline)
         .onAppear {
