+++ Flipcash/Core/Screens/Onramp/EnterEmailScreen.swift
@@ -59,7 +59,7 @@ struct EnterEmailScreen: View {
             .padding(20)
             .foregroundStyle(.textMain)
         }
-        .trackedDialog(item: $viewModel.dialogItem)
+        .dialog(item: $viewModel.dialogItem)
         .navigationTitle("Verify Email")
         .toolbarTitleDisplayMode(.inline)
         .onAppear {
